requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
beautifulsoup4==4.11.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.0.1
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-cors-headers==4.0.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
idna==3.4
numpy==1.24.1
pandas==2.0.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
tzdata==2023.3
urllib3==1.26.14
wincertstore==0.2
